worker_Server/requirements.txt
requests>=2.25.1
pymongo>=3.11.0
python-dotenv>=0.19.0
transformers>=4.30.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.65.0